+++ src/routes/LPOptimisers/Vaults/VaultsTable/MellowPosition/MellowPosition.styled.tsx
@@ -3,8 +3,8 @@ import Button from '@mui/material/Button';
 import Skeleton from '@mui/material/Skeleton';
 import { styled } from '@mui/material/styles';
 
-import { Typography } from '../../../../../../components/atomic/Typography/Typography';
-import colors from '../../../../../../theme/colors';
+import { Typography } from '../../../../../components/atomic/Typography/Typography';
+import colors from '../../../../../theme/colors';
 
 export const MellowPositionSkeleton = styled(Skeleton)`
   width: 100%;
