+++ src/routes/LPOptimisers/Vaults/VaultsTable/MellowPosition/MellowPosition.tsx
@@ -1,7 +1,7 @@
 import isUndefined from 'lodash.isundefined';
 import React from 'react';
 
-import { formatCurrency } from '../../../../../../utilities/number';
+import { formatCurrency } from '../../../../../utilities/number';
 import {
   DepositButton,
   DepositTypography,
