+++ src/routes/LPOptimisers/Vaults/VaultsTable/Tag/Tag.styled.tsx
@@ -1,8 +1,8 @@
 import Box from '@mui/material/Box';
 import { styled } from '@mui/material/styles';
 
-import { Typography } from '../../../../../../components/atomic/Typography/Typography';
-import { colors } from '../../../../../../theme';
+import { Typography } from '../../../../../components/atomic/Typography/Typography';
+import { colors } from '../../../../../theme';
 
 export const TagContainer = styled(Box)`
   display: flex;
